pyproject.toml
[project]
name = "induction-hardening-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.15.1",
    "jupyterlab>=4.5.0",
    "matplotlib>=3.10.7",
    "neuraloperator>=2.0.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
